requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
idna==3.7
load-dotenv==0.1.0
multidict==6.0.5
outcome==1.3.0.post0
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.0
trio-websocket==0.11.1
twilio==9.2.3
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
